refactored to use git bisect replay

+++ scripts/bisection/bisect.sh
@@ -1,42 +1,54 @@
 #!/bin/bash
-set -x
 
 test=$1
 count=$2
 duration_mins=$3
 
-. ./scripts/bisection/bisect-util.sh
+SCRIPT_DIR=$(dirname "$0")
+. "$SCRIPT_DIR"/bisect-util.sh
 
-BISECT_DIR="$(clean_test_name "$test")"
-LOG_NAME="$BISECT_DIR/bisection.log"
-CONF_NAME="$BISECT_DIR/config.json"
 CURRENT_HASH="$(current_hash)"
 
-#echo "dummy ops"
-#read opsPerSec
+#first lets check our saved results
+opsPerSec=$(get_hash_result "$CURRENT_HASH")
+
+case $opsPerSec in
+  USER_GOOD)
+    exit 0
+    ;;
+  USER_BAD)
+    exit 1
+    ;;
+  USER_SKIP)
+    exit 128
+    ;;
+  *)
+    ;;
+esac
+
 #./pkg/cmd/roachtest/roachstress.sh -b -c 2 "^$test\$" -- --parallelism 1 -c "$cluster" --debug-reuse
 
-#trap 'echo error on line $LINENO bisecting $CURRENT_HASH. aborting; git reset --hard; exit 128' ERR
+#if opsPerSec was not previously saved, we calculate it by building and running
+if [[ -z "$opsPerSec" ]]; then
+  build_sha "$CURRENT_HASH" "$duration_mins" &> "$BISECT_DIR/$CURRENT_HASH-build.log"
+  stress_sha "$CURRENT_HASH" "$test" "$count" &> "$BISECT_DIR/$CURRENT_HASH-run.log"
 
-build_sha "$CURRENT_HASH" "$duration_mins" &> "$BISECT_DIR/$CURRENT_HASH-build.log"
-stress_sha "$CURRENT_HASH" "$test" "$count" &> "$BISECT_DIR/$CURRENT_HASH-run.log"
+  git reset --hard
 
-git reset --hard
+  #looks under the artifacts directory for the current runs stats to calculate the average/second
+  opsPerSec=$(calc_avg_ops "artifacts/$CURRENT_HASH*/$test/run_*/*.perf/stats.json")
+  set_hash_result "$CURRENT_HASH" "$opsPerSec"
+fi
 
-#looks under the artifacts directory for the current runs stats to calculate the average/second
-opsPerSec=$(calc_avg_ops "artifacts/$CURRENT_HASH*/$test/run_*/*.perf/stats.json")
-
-goodThreshold=$(get_conf_val ".goodThreshold" "$CONF_NAME")
-badThreshold=$(get_conf_val ".badThreshold" "$CONF_NAME")
+goodThreshold=$(get_conf_val ".goodThreshold")
+badThreshold=$(get_conf_val ".badThreshold")
 
 if [ -n "$goodThreshold" ] && [[ opsPerSec -ge goodThreshold ]]; then
   log "[$CURRENT_HASH] Average ops/s: [$opsPerSec]. Auto marked as good." "$LOG_NAME"
-  set_conf_val "goodHash" "$CURRENT_HASH" "$CONF_NAME"
   exit 0;
 elif [ -n "$badThreshold" ] && [[ opsPerSec -le badThreshold ]]; then
   log "[$CURRENT_HASH] Average ops/s: [$opsPerSec]. Auto marked as bad." "$LOG_NAME"
-  set_conf_val "badHash" "$CURRENT_HASH" "$CONF_NAME"
   exit 1;
 else
-  prompt_user "$CURRENT_HASH" "$opsPerSec" "$CONF_NAME" "$LOG_NAME"
+  prompt_user "$CURRENT_HASH" "$opsPerSec"
 fi
